added updateWordPressImage function
