+++ src/lib/dev-environment/dev-environment-cli.js
@@ -7,14 +7,12 @@
  * External dependencies
  */
 import chalk from 'chalk';
-import fetch from 'node-fetch';
 import formatters from 'lando/lib/formatters';
 import { prompt, Confirm, Select } from 'enquirer';
 import debugLib from 'debug';
 import fs from 'fs';
 import path from 'path';
 import os from 'os';
-import xdgBasedir from 'xdg-basedir';
 
 /**
  * Internal dependencies
@@ -26,11 +24,9 @@ import {
 	DEV_ENVIRONMENT_PROMPT_INTRO,
 	DEV_ENVIRONMENT_COMPONENTS,
 	DEV_ENVIRONMENT_NOT_FOUND,
-	DEV_ENVIRONMENT_RAW_GITHUB_HOST,
-	DEV_ENVIRONMENT_WORDPRESS_VERSIONS_URI,
-	DEV_ENVIRONMENT_WORDPRESS_CACHE_KEY,
 } from '../constants/dev-environment';
 import { InstanceOptions, EnvironmentNameOptions, InstanceData } from './types';
+import { getVersionList } from './dev-environment-core';
 
 const debug = debugLib( '@automattic/vip:bin:dev-environment' );
 
@@ -348,63 +344,9 @@ export function addDevEnvConfigurationOptions( command ) {
 }
 
 /**
- * Makes a web call to raw.githubusercontent.com
- */
-async function fetchVersionList() {
-	const url = `https://${ DEV_ENVIRONMENT_RAW_GITHUB_HOST }${ DEV_ENVIRONMENT_WORDPRESS_VERSIONS_URI }`;
-	return fetch( url ).then( res => res.text() );
-}
-
-/**
- * Uses a cache file to keep the version list in tow until it is ultimately outdated
- */
-async function getVersionList() {
-	let res;
-	const local = xdgBasedir.data || os.tmpdir();
-	const cacheTtl = 86400; // number of seconds that the cache can be considered active.
-	const cacheFile = path.join( local, 'vip', DEV_ENVIRONMENT_WORDPRESS_CACHE_KEY );
-
-	// Try to retrieve the file from cache or cache it if invalid
-	try {
-		// If the cache doesn't exist, create it
-		if ( ! fs.existsSync( cacheFile ) ) {
-			res = await fetchVersionList();
-			fs.writeFileSync( cacheFile, res );
-		}
-
-		// Last modified
-		const stats = fs.statSync( cacheFile );
-		debug( `WordPress Version List cache last modified: ${ stats.mtime }` );
-
-		// If the cache is expired, fetch the list again and cache it
-		const expire = new Date( stats.mtime );
-		expire.setSeconds( expire.getSeconds() + cacheTtl );
-
-		if ( +new Date > expire ) {
-			debug( `WordPress Version List cache is expired: ${ expire }` );
-			res = await fetchVersionList();
-			fs.writeFileSync( cacheFile, res );
-		}
-	} catch ( err ) {
-		// Soft error handling here, since it's still possible to use a previously cached file.
-		console.log( chalk.yellow( 'fetchWordPressVersionList failed to retrieve an updated version list' ) );
-		debug( err );
-	}
-
-	// Try to parse the cached file if it exists
-	// if not, something worse than a failed request happend; bail.
-	try {
-		return JSON.parse( fs.readFileSync( cacheFile ) );
-	} catch ( err ) {
-		debug( err );
-		return [];
-	}
-}
-
-/**
  * Provides the list of tag choices for selection
  */
-async function getTagChoices() {
+export async function getTagChoices() {
 	const tagChoices = [];
 	let tagFormatted, prerelease, mapping;
 	const versions = await getVersionList();
